Added cache [more in desc]
Added a way for images to be stored inside the database so that they do not need to be re-rendered. This takes the time to serve a webp from 400ms to 20ms. Speed. Im also adding a new type of object called a history or audit log, still working on the name.